requirements-common.txt
numpy
pillow
tqdm
pyyaml
torch
torchvision
transformers
sacrebleu
rouge-score
nltk
bert-score
sentence-transformers
scikit-learn
